Renamed packages
"hectoserver" -> "ns" + "ns/lookup".
