+++ ns/lookup/conn_test.go
@@ -1,4 +1,4 @@
-package hectoserver
+package lookup
 
 import (
 	"context"
