+++ ns/request.go
@@ -1,20 +1,19 @@
-package hectoserver
+package ns
 
 import (
 	"bufio"
 	"bytes"
-	"context"
 	"encoding/base64"
 	"fmt"
 	"io"
 	"io/ioutil"
 	"net"
 	"net/http"
+	"strconv"
+	"strings"
 	"time"
 
 	"github.com/miekg/dns"
-	"github.com/pkg/errors"
-	"github.com/rs/zerolog"
 )
 
 const (
@@ -24,6 +23,10 @@ const (
 
 // Error represents a server error.
 type Error struct{ err string }
+
+func NewError(text string) Error {
+	return Error{err: text}
+}
 
 // Error returns string representation of the error.
 func (e Error) Error() string { return "hectodns: " + e.err }
@@ -265,90 +268,91 @@ func (r *Response) Redirect() *Response {
 	return r
 }
 
-type Handler interface {
-	Handle(context.Context, *Request) (*Response, error)
+type ForwardedHeader struct {
+	Proto string
+	Host  string
+	By    net.TCPAddr
+	For   []net.TCPAddr
 }
 
-// HandlerFunc is a function adapter to allow use of ordinary functions
-// as request handlers. If f is a function with the appropriate signature,
-// HandlerFunc(f) is a Handler that calls f.
-type HandlerFunc func(context.Context, *Request) (*Response, error)
+// ParseForwarded parses a Forwarded header and returns all passed
+// directives within a forwarded map.
+func ParseForwarded(header string) (fw ForwardedHeader, err error) {
+	var (
+		raw_dirs = strings.Split(header, ";")
+		dirs     = make([]string, 0, len(raw_dirs))
+	)
 
-// Handle implements Handler interface, calls f(ctx, req)
-func (fn HandlerFunc) Handle(ctx context.Context, req *Request) (*Response, error) {
-	return fn(ctx, req)
-}
-
-// timeoutHandler is a handler that limits the handler execution by the
-// specified timeout. Wrapped handler must implement context cancellation.
-type timeoutHandler struct {
-	timeout time.Duration
-	Handler
-}
-
-// Handle implements Handler interface. It adds a timeout to the passed
-// context and then executes a wrapped handler instance.
-func (h timeoutHandler) Handle(ctx context.Context, req *Request) (*Response, error) {
-	ctx, cancel := context.WithTimeout(ctx, h.timeout)
-	defer cancel()
-
-	return h.Handler.Handle(ctx, req)
-}
-
-// MultiHandler creates a handler that executes all passed handlers one
-// by one until one of them successfully processes the request.
-//
-// If a listed handler returns an error, the overall handle operation
-// stops and returns the error; it does not continue down the list.
-func MultiHandler(hh ...Handler) Handler {
-	return &multiHandler{hh}
-}
-
-type multiHandler struct {
-	handlers []Handler
-}
-
-// Handle implements Handler interface. Method sequentially executes
-// handlers until the successful response is received.
-//
-// Headers from previous response are copied to the next request, so
-// the handlers could influence behaviour of next handlers.
-func (mh *multiHandler) Handle(ctx context.Context, req *Request) (*Response, error) {
-	log := zerolog.Ctx(ctx).With().Uint16("id", req.Body.Id).Logger()
-	log.Debug().Msg("received request")
-
-	if log.GetLevel() <= zerolog.DebugLevel {
-		b, _ := req.Body.Pack()
-		b64 := base64.RawURLEncoding.EncodeToString(b)
-		log.Debug().Str("b64", b64).Msg("request received")
+	// Each directive can represent a list of values, like it happens for
+	// a directive "for":
+	//
+	// 	Forwarded: for=192.168.0.1,for=192.168.0.2;by=192.168.0.254
+	for _, directive := range raw_dirs {
+		dirs = append(dirs, strings.Split(directive, ",")...)
 	}
 
-	// Bypass the logger though context of the handler.
-	ctx = log.WithContext(ctx)
+	for _, directive := range dirs {
+		// Each directive reprsents a key-value pair: by=1.2.3.4
+		kv := strings.SplitN(strings.Trim(directive, " "), "=", 2)
+		if len(kv) != 2 {
+			return ForwardedHeader{}, err
+		}
 
-	for no, h := range mh.handlers {
-		log.Debug().Msgf("passed request to %d resolver", no)
-		resp, err := h.Handle(ctx, req)
+		key := strings.ToLower(kv[0])
+		val := kv[1]
 
+		// For directives that does not support lists (proto, host, by),
+		// simply override the value when a new identifier appears.
+		switch key {
+		case "proto":
+			fw.Proto = val
+		case "host":
+			fw.Host = val
+		case "by":
+			ip, port, err := parseHostPort(val)
+			if err != nil {
+				return ForwardedHeader{}, err
+			}
+			fw.By = net.TCPAddr{IP: ip, Port: port}
+		case "for":
+			ip, port, err := parseHostPort(val)
+			if err != nil {
+				return ForwardedHeader{}, err
+			}
+			fw.For = append(fw.For, net.TCPAddr{IP: ip, Port: port})
+		}
+	}
+
+	return
+}
+
+func parseHostPort(s string) (ip net.IP, port int, err error) {
+	if strings.HasPrefix(s, `"`) && strings.HasSuffix(s, `"`) {
+		s = strings.TrimLeft(s, `"`)
+		s = strings.TrimRight(s, `"`)
+	}
+
+	portIndex := strings.LastIndex(s, ":")
+
+	if portIndex > 0 {
+		i, err := strconv.ParseInt(s[portIndex+1:], 10, 0)
 		if err != nil {
-			log.Debug().Msgf("error from %d resolver, %s", no, err)
-			return nil, err
+			return nil, 0, err
 		}
 
-		log.Debug().
-			Int("status", resp.StatusCode).
-			Int("rcode", resp.Body.Rcode).
-			Msgf("received reply")
-
-		if resp.StatusCode == http.StatusOK {
-			log.Debug().Msgf("sending reply")
-			return resp, nil
-		}
-
-		// Bypass headers to the next handlers, so the handlers could
-		// exchange intermediate information.
-		req.Header = resp.Header
+		port = int(i)
+		s = s[:portIndex]
 	}
 
-	return nil, errors.New("no handlers left")
+	if strings.HasPrefix(s, `[`) && strings.HasSuffix(s, `]`) {
+		s = strings.TrimLeft(s, `[`)
+		s = strings.TrimRight(s, `]`)
+	}
+
+	ip = net.ParseIP(s)
+	if ip == nil {
+		return ip, port, &net.ParseError{Type: "IP address", Text: s}
+	}
+
+	return
 }
